requirements.txt
python-telegram-bot==13.15
Pillow==9.2.0
